services/career-advisor-service/requirements.txt
docx==0.2.4
fastapi==0.115.12
httpx==0.28.1
numpy==2.2.4
openai==1.72.0
pinecone==6.0.2
prometheus_client==0.21.1
pydantic==2.11.3
pydantic_settings==2.8.1
PyPDF2==3.0.1
python-dotenv==1.1.0
python_docx==1.1.2
redis==5.2.1
sentence_transformers==4.0.2
SQLAlchemy==2.0.38
tenacity==9.0.0
uvicorn==0.34.0
psycopg2-binary==2.9.10
alembic==1.15.2
python-multipart==0.0.20